# apps/server/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "DDAS API",
    "version": "2.0.0-alpha.0",
    "description": "Dynamic Delegation of Authority System — appetite-constrained authority routing for humans and AI agents."
  },
...
  "paths": {
...
    "/api/v1/audit/events": {
      "get": {
        "tags": [
          "audit"
        ],
        "parameters": [
          {
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0
            },
            "in": "query",
            "name": "after",
            "required": false
          },
          {
            "schema": {
              "type": "integer",
              "exclusiveMinimum": true,
              "minimum": 0,
              "maximum": 500,
              "default": 100
            },
            "in": "query",
            "name": "limit",
            "required": false
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "type",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "seq": {
                        "type": "number"
                      },
                      "occurredAt": {
                        "type": "string"
                      },
                      "actor": {},
                      "type": {
                        "type": "string"
                      },
                      "entity": {},
                      "payload": {},
                      "eventHash": {
                        "type": "string"
                      }
                    },
                    "required": [
                      "seq",
                      "occurredAt",
                      "type",
                      "eventHash"
                    ],
                    "additionalProperties": false
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "/"
    }
  ]
}